Make memoryfs usage consistent across internal utilities
While extractor allowed a configuration object with a targetFs attribute which can have a memoryFs instance,
the availability of this configuration is not useful unless the rest of the internal functions (cleaner, compiler etc.)
also facilitate passing a targetFs configuration attribute and use it internally.

After this change, only the middlewares would be aware of the global memoryfs singleton and the internal functions
would rely on the passed memoryFs instance.